Move script/util/test-utils.js to test/util/utils.js
