+++ test/integration/03-DR-removes-the-seller-from-allowed-list.js
@@ -11,7 +11,7 @@ const { deployProtocolHandlerFacets } = require("../../scripts/util/deploy-proto
 const { deployProtocolConfigFacet } = require("../../scripts/util/deploy-protocol-config-facet.js");
 const { deployProtocolClients } = require("../../scripts/util/deploy-protocol-clients");
 const { deployMockTokens } = require("../../scripts/util/deploy-mock-tokens");
-const { setNextBlockTimestamp, applyPercentage } = require("../../scripts/util/test-utils.js");
+const { setNextBlockTimestamp, applyPercentage } = require("../util/utils.js");
 const { oneWeek, oneMonth } = require("../utils/constants");
 const {
   mockOffer,
