+++ test/integration/04-DR-removes-fees.js
@@ -21,11 +21,7 @@ const { deployProtocolHandlerFacets } = require("../../scripts/util/deploy-proto
 const { deployProtocolConfigFacet } = require("../../scripts/util/deploy-protocol-config-facet.js");
 const { deployProtocolClients } = require("../../scripts/util/deploy-protocol-clients");
 const { oneMonth, oneWeek } = require("../utils/constants");
-const {
-  setNextBlockTimestamp,
-  calculateContractAddress,
-  applyPercentage,
-} = require("../../scripts/util/test-utils.js");
+const { setNextBlockTimestamp, calculateContractAddress, applyPercentage } = require("../util/utils.js");
 
 /**
  *  Integration test case - exchange and offer operations should remain possible even when token fees are removed from the DR fee list
