+++ test/protocol/BuyerHandlerTest.js
@@ -11,7 +11,7 @@ const { deployProtocolDiamond } = require("../../scripts/util/deploy-protocol-di
 const { deployProtocolHandlerFacets } = require("../../scripts/util/deploy-protocol-handler-facets.js");
 const { deployProtocolConfigFacet } = require("../../scripts/util/deploy-protocol-config-facet.js");
 const { deployProtocolClients } = require("../../scripts/util/deploy-protocol-clients");
-const { calculateContractAddress } = require("../../scripts/util/test-utils.js");
+const { calculateContractAddress } = require("../util/utils.js");
 const { oneWeek, oneMonth } = require("../utils/constants");
 const {
   mockOffer,
